extract component that displays selected guests+intolerances
